+++ CHANGES.md
@@ -7,3 +7,4 @@
 ## 2023-09-xx 0.0.0
 - Initial thing, proof-of-concept
 - Add software tests
+- CLI: Add `mlflow-cratedb cratedb --version` command